+++ src/components/entity/enemies/Octorok.java
@@ -107,8 +107,8 @@ public class Octorok extends Enemy
 
 	public void draw(Graphics2D g2d)
 	{
-		drawX = x - overWorld.getCameraX();
-		drawY = y - overWorld.getCameraY();
+		drawX = (int) x - overWorld.getCameraX();
+		drawY = (int) y - overWorld.getCameraY();
 
 		if(pellet != null) pellet.draw(g2d);
